+++ tests/testthat/test-fit_r_light.R
@@ -1,60 +1,75 @@
 library(testthat)
 library(photosynthesis)
-context("Fitting Rlight")
+context("Estimating light respiration")
 
-# REVISIT AFTER fit_r_light2() is done
-df <- data.frame(
-  A_net = c(
+df = data.frame(
+  .A = c(
     seq(from = -3, to = -1.5, by = 0.5),
     seq(from = -1.25, to = 1, by = 0.25)
   ),
-  PPFD = c(seq(from = 0, to = 130, by = 10))
+  .Q = c(seq(from = 0, to = 130, by = 10))
 )
 
-# model <- fit_r_light_kok(df)
+fit = fit_photosynthesis(
+  .data = df,
+  .photo_fun = "r_light",
+  .model = "kok_1956",
+  Q_lower = 0,
+  Q_upper = 130
+)
 
-# test_that("Outputs", {
-#   expect_is(object = model, class = "numeric")
-# })
+test_that("kok_1956 output", {
+  expect_is(object = fit, class = "lm")
+})
 
-df <- data.frame(
-  A_net = c(
+df = data.frame(
+  .A = c(
     seq(from = 1, to = 5, by = 1),
     seq(from = 0.5, to = 2.5, by = 0.5),
     seq(from = 0.25, to = 1.25, by = 0.25)
   ),
-  C_i = c(
+  .C = c(
     50, 100, 150, 200, 250,
     75, 125, 175, 225, 275,
     100, 150, 200, 250, 299
   ),
-  PPFD = c(
+  .Q = c(
     rep(1500, 5),
     rep(750, 5),
     rep(375, 5)
   )
 )
 
-# model <- fit_r_light_WalkerOrt(df)
+fit = fit_photosynthesis(
+  .data = df,
+  .photo_fun = "r_light",
+  .model = "walker_ort_2015",
+  C_upper = 300,
+  Q_levels = c(1500, 750, 375)
+)
 
-# test_that("Outputs", {
-#   expect_is(object = model[1], class = "list")
-#   expect_is(object = model[2], class = "list")
-#   expect_is(object = model[[3]], class = "data.frame")
-#   expect_length(object = model, 3)
-# })
+test_that("walker_ort_2015 output", {
+  expect_is(object = fit, class = "lm")
+})
 
-df <- data.frame(
-  A_net = c(
+
+df = data.frame(
+  .A = c(
     seq(from = -3, to = -1.5, by = 0.5),
     seq(from = -1.25, to = 1, by = 0.25)
   ),
-  PPFD = c(seq(from = 0, to = 130, by = 10)),
-  phi_PSII = c(seq(from = 0.74, to = 0.61, by = -0.01))
+  .Q = c(seq(from = 0, to = 130, by = 10)),
+  .phiPSII = c(seq(from = 0.74, to = 0.61, by = -0.01))
 )
 
-# model <- fit_r_light_yin(df)
+fit = fit_photosynthesis(
+  .data = df,
+  .photo_fun = "r_light",
+  .model = "yin_etal_2011",
+  Q_lower = 0,
+  Q_upper = 130
+)
 
-# test_that("Outputs", {
-#   expect_is(object = model, class = "numeric")
-# })
+test_that("yin_etal_2011 output", {
+  expect_is(object = fit, class = "lm")
+})